+++ packages/amplify-e2e-tests/src/__tests__/apigw.test.ts
@@ -1,6 +1,7 @@
 import {
   addRestApi,
   createNewProjectDir,
+  get,
   initJSProjectWithProfile,
   deleteProject,
   deleteProjectDir,
@@ -26,12 +27,16 @@ afterAll(async () => {
 
 describe('API Gateway e2e tests', () => {
   it('adds multiple rest apis and pushes', async () => {
-    await addRestApi(projRoot, {});
+    const firstRestApi = `firstE2eRestApi${shortId}`;
+    const secondRestApi = `secondE2eRestApi${shortId}`;
+
+    await addRestApi(projRoot, { apiName: firstRestApi });
     await amplifyPushAuth(projRoot);
     await addAuthWithGroupsAndAdminAPI(projRoot); // Groups: Admins, Users
     await amplifyPushAuth(projRoot);
-    await addRestApi(projRoot, { isFirstRestApi: false, path: '/foo', projectContainsFunctions: true });
+    await addRestApi(projRoot, { isFirstRestApi: false, path: '/foo', projectContainsFunctions: true }); // Add a path
     await addRestApi(projRoot, {
+      apiName: secondRestApi,
       isFirstRestApi: false,
       restrictAccess: true,
       allowGuestUsers: true,
@@ -39,8 +44,29 @@ describe('API Gateway e2e tests', () => {
       projectContainsFunctions: true,
     });
     await amplifyPushAuth(projRoot); // Pushes multiple rest api updates
+
     const projMeta = getProjectMeta(projRoot);
     expect(projMeta).toBeDefined();
     expect(projMeta.api).toBeDefined();
+    expect(projMeta.api.AdminQueries).toBeDefined();
+    expect(projMeta.api[firstRestApi]).toBeDefined();
+    expect(projMeta.api[secondRestApi]).toBeDefined();
+
+    const firstRootUrl = projMeta.api[firstRestApi].output?.RootUrl;
+    const secondRootUrl = projMeta.api[secondRestApi].output?.RootUrl;
+    expect(firstRootUrl).toBeDefined();
+    expect(secondRootUrl).toBeDefined();
+
+    const firstItemsResponse = await get(`${firstRootUrl}/items`);
+    const fooResponse = await get(`${firstRootUrl}/foo`);
+    const secondItemsResponse = await get(`${secondRootUrl}/items`);
+
+    const firstItemsResJson = await firstItemsResponse.json();
+    const fooResJson = await fooResponse.json();
+    const secondItemsResJson = await secondItemsResponse.json();
+
+    expect(firstItemsResJson).toEqual({ success: 'get call succeed!', url: '/items' });
+    expect(fooResJson).toEqual({ success: 'get call succeed!', url: '/foo' });
+    expect(secondItemsResJson).toEqual({ message: 'Missing Authentication Token' }); // Restricted API
   });
 });