+++ packages/amplify-headless-interface/src/interface/geo/add.ts
@@ -12,24 +12,62 @@
   serviceConfiguration: GeoServiceConfiguration;
 }
 
-export interface GeoServiceConfiguration {
+/**
+ * Defines AWS Location Service parameters.
+ */
+export type GeoServiceConfiguration = BaseGeoServiceConfiguration & MapConfiguration;
+
+/**
+ * Configuration that applies to all geo service configuration.
+ */
+export interface BaseGeoServiceConfiguration {
+  /**
+   * The service name of the resource provider.
+   */
   serviceName: string;
+  /**
+   * The name of the map that will be created.
+   */
   name: string;
+  /**
+   * The access policy for geo resources.
+   */
   accessType: AccessType;
+  /**
+   * The pricing plan for amazon location service.
+   */
   pricingPlan: PricingPlan;
+  /**
+   * Whether the geo resource added is set to default.
+   */
   setAsDefault: boolean;
 }
-export interface MapConfiguration extends GeoServiceConfiguration {
+/**
+ * Specifies configuration for map.
+ */
+export interface MapConfiguration {
+  /**
+   * The service name of the resource provider.
+   */
   serviceName: "Map";
+  /**
+   * The map style type.
+   */
   mapStyle: MapStyle;
 }
 
+/**
+ * Pricing plan of Amazon Location Service.
+ */
 export enum PricingPlan {
   RequestBasedUsage = "RequestBasedUsage",
   MobileAssetTracking = "MobileAssetTracking",
   MobileAssetManagement = "MobileAssetManagement"
 }
 
+/**
+ * Definition of access type
+ */
 export enum AccessType {
   AuthorizedUsers = "AuthorizedUsers",
   AuthorizedAndGuestUsers = "AuthorizedAndGuestUsers"