+++ packages/graphql-key-transformer/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.11](https://github.com/aws-amplify/amplify-cli/compare/graphql-key-transformer@3.2.10...graphql-key-transformer@3.2.11) (2021-12-21)
+
+**Note:** Version bump only for package graphql-key-transformer
+
+
+
+
+
+## [3.2.10](https://github.com/aws-amplify/amplify-cli/compare/graphql-key-transformer@3.2.9...graphql-key-transformer@3.2.10) (2021-12-17)
+
+**Note:** Version bump only for package graphql-key-transformer
+
+
+
+
+
 ## [3.2.9](https://github.com/aws-amplify/amplify-cli/compare/graphql-key-transformer@3.2.8...graphql-key-transformer@3.2.9) (2021-12-03)
 
 **Note:** Version bump only for package graphql-key-transformer